fix a bug of nand_block.c in nand_rq_map_sl()

+++ drivers/nand/inc/lpartition.h
@@ -20,7 +20,7 @@ struct _LPartition {
     PartContext* pc;
 	
 	int hwsector;
-	unsigned short segmentsize;
+	unsigned int segmentsize;
 };
 
 #endif
